Silence misleading exception print in setQtContextDelegate
As I understand it, the call to getDeclaredMethod seems to be just there
to check if a given method is declared under before adding it to the
delegates list, and that's why the getDeclaredMethod is not returning
anything, so we could treat it as a failed check and not print anything.

This is an amendment to 80f7494e8a9f9a70e3b53833a098d74d8c2331d9 which
added this print.

Change-Id: I5f69ed5b4fa655da53ac7fba20d4e07acc75607a
Reviewed-by: Andy Shaw <andy.shaw@qt.io>
(cherry picked from commit eaac5bae9838a73d0d70004bcb37b97635d1f2d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>